ENH: Output differentials to TSV instead of CSV
This adjusts the README accordingly.

Worth noting that the Q2 code creates a differentials QZA file
in which the underyling data frame is named differential.tsv (note the
lack of trailing "s" in "differential"). I'll see if I can change
that, in order to make the Q2 and standalone outputs more consistent.