ENH: Alter Q2 output: differential->differentials
This should make the Q2 and standalone outputs match, which will
alleviate confusion.

There are still a few places where the single "Differential" instead of
"Differentials" is used (e.g. the differential_prior argument,
FeatureData[Differential]), but at least this way the output (and
related params -- that is, --o-differentials) is consistent. Not
sure if those should also be changed.