DOC: Make Tensorboard instructions more visible
This should make it clearer that you need to actually use this
for validating that your model is fitting properly.

Even though this is an important step, I've seen many people download
Songbird, run it once, and just ignore the tensorboard stuff because
it doesn't look important. (I was also guilty of this for a while, so
something something glass houses.) Making this more visible will
help indicate to people how to use this tool properly.