ENH: Remove --i-feature-table for summarize-single
The fact that "n" is now optional in _summarize() means that there's
ostensibly a weird case where _summarize() could be called *with*
baseline stats but without n. So I made _summarize() throw an error
in that case (should never happen but might as well be defensive),
and also added a corresponding test case.

This closes #76.