REL: Bump up minimum tensorflow versions
ideally we should be using 1.15.2 for both conda and pip, but
conda-forge doesn't seem to have 1.15.2 yet. hence the slightly
different minimum versions. (... not sure if this will work since
conda-forge's latest tf v1 build seems to be for 1.13? we'll see
on travis.)

i have verified on an ubuntu computer that the tests pass with tf
1.15.2 (installed via pip)